test: migrate a bunch of CC tests to AVA
