+++ packages/zwave-js/src/lib/test/cc/ManufacturerSpecificCC.test.ava.ts
@@ -1,0 +1,51 @@
+import {
+	ManufacturerSpecificCCGet,
+	ManufacturerSpecificCCReport,
+	ManufacturerSpecificCommand,
+} from "@zwave-js/cc";
+import { CommandClasses } from "@zwave-js/core";
+import { createTestingHost } from "@zwave-js/host";
+import test from "ava";
+
+const host = createTestingHost();
+
+function buildCCBuffer(payload: Buffer): Buffer {
+	return Buffer.concat([
+		Buffer.from([
+			CommandClasses["Manufacturer Specific"], // CC
+		]),
+		payload,
+	]);
+}
+
+test("the Get command should serialize correctly", (t) => {
+	const cc = new ManufacturerSpecificCCGet(host, { nodeId: 1 });
+	const expected = buildCCBuffer(
+		Buffer.from([
+			ManufacturerSpecificCommand.Get, // CC Command
+		]),
+	);
+	t.deepEqual(cc.serialize(), expected);
+});
+
+test("the Report command (v1) should be deserialized correctly", (t) => {
+	const ccData = buildCCBuffer(
+		Buffer.from([
+			ManufacturerSpecificCommand.Report, // CC Command
+			0x01,
+			0x02,
+			0x03,
+			0x04,
+			0x05,
+			0x06,
+		]),
+	);
+	const cc = new ManufacturerSpecificCCReport(host, {
+		nodeId: 2,
+		data: ccData,
+	});
+
+	t.is(cc.manufacturerId, 0x0102);
+	t.is(cc.productType, 0x0304);
+	t.is(cc.productId, 0x0506);
+});
